Move MoshiElement parse and write to own module
